limiting access to requests / added superuser rights

+++ src/category/category.controller.ts
@@ -7,6 +7,8 @@ import {
   Param,
   Post,
   Query,
+  Req,
+  UseGuards,
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 
@@ -16,6 +18,9 @@ import { checkIsObjectIdValid } from 'src/utils/checkIsObjectIdValid';
 import { CategoryService } from './category.service';
 import { ImagesService } from 'src/images/images.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
+import { JwtGuard } from '../guard/jwt.guard';
+import { Users } from '../users/schemas/users.schema';
+import { checkIfUserHasPermissionToChangeRestaurant } from '../utils/checkIfUserHasPermissionToChangeRestaurant';
 
 @Controller('category')
 export class CategoryController {
@@ -44,14 +49,25 @@ export class CategoryController {
       : this.menuService.findByCategoryId(id);
   }
 
+  @UseGuards(JwtGuard)
   @Post(':id/menu-item')
-  async createMenuItem(@Body() dto: MenuItemsDto, @Param('id') id: string) {
+  async createMenuItem(
+    @Body() dto: MenuItemsDto,
+    @Param('id') id: string,
+    @Req() req,
+  ) {
+    const user: Users = req.user;
     await checkIsObjectIdValid(id);
 
     const category = await this.categoryService.findById(id);
     if (!category) {
       throw new NotFoundException();
     }
+
+    await checkIfUserHasPermissionToChangeRestaurant(
+      user,
+      category.restaurant._id,
+    );
 
     const isPictureExists = await this.imagesService.checkFileForExistence(
       dto.pictureUrl,
@@ -63,20 +79,33 @@ export class CategoryController {
     return this.menuService.create(dto, category, dto.pictureUrl);
   }
 
+  @UseGuards(JwtGuard)
   @Delete(':id')
-  async removeCategory(@Param('id') id: string) {
+  async removeCategory(@Param('id') id: string, @Req() req) {
+    const user: Users = req.user;
     await checkIsObjectIdValid(id);
 
     const category = await this.categoryService.findById(id);
     if (!category) {
       throw new NotFoundException();
     }
+
+    await checkIfUserHasPermissionToChangeRestaurant(
+      user,
+      category.restaurant._id,
+    );
 
     return this.categoryService.removeCategory(id);
   }
 
+  @UseGuards(JwtGuard)
   @Post(':id/update')
-  async updateById(@Body() dto: CreateCategoryDto, @Param('id') id: string) {
+  async updateById(
+    @Body() dto: CreateCategoryDto,
+    @Param('id') id: string,
+    @Req() req,
+  ) {
+    const user: Users = req.user;
     await checkIsObjectIdValid(id);
 
     const category = await this.categoryService.findById(id);
@@ -84,15 +113,23 @@ export class CategoryController {
       throw new NotFoundException();
     }
 
+    await checkIfUserHasPermissionToChangeRestaurant(
+      user,
+      category.restaurant._id,
+    );
+
     return this.categoryService.updateById(id, dto);
   }
 
+  @UseGuards(JwtGuard)
   @Post(':categoryId1/swap/:categoryId2')
   async changeItemNumber(
     @Body() dto: CreateCategoryDto,
     @Param('categoryId1') categoryId1: string,
     @Param('categoryId2') categoryId2: string,
+    @Req() req,
   ) {
+    const user: Users = req.user;
     await checkIsObjectIdValid(categoryId1);
     await checkIsObjectIdValid(categoryId2);
 
@@ -102,6 +139,11 @@ export class CategoryController {
         `Categories with id(${category1}) does not exist`,
       );
     }
+
+    await checkIfUserHasPermissionToChangeRestaurant(
+      user,
+      category1.restaurant._id,
+    );
 
     const category2 = await this.categoryService.findById(categoryId2);
     if (!category2) {
